utilisation editor, recherche utilisation state areaValue
